Proper default value for name commands.
For name_register and name_update, we should use "{}" as the default value
if none is provided explicitly.  This is the simplest possible value that can
be used in Xaya, i.e. what one should use to just register a name
without any specific data yet, or e.g. just update it for a transfer.